requirements.txt
pandas 
mlflow
notebook
numpy
scikit-learn
matplotlib
python-box
pyYAML
tqdm
ensure
joblib
types-PyYAML
Flask
Flask-Cors

ensure
box



#-e .